Move `filters.ts`
